requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
attrs==23.1.0
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
defusedxml==0.7.1
distlib==0.3.7
distro==1.8.0
filelock==3.13.1
Flask==3.0.0
frozenlist==1.4.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpie==3.2.2
httpx==0.25.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
openai==0.27.1
packaging==23.2
platformdirs==3.11.0
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.6.0
setuptools==68.2.2
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
virtualenv==20.24.6
waitress==2.1.2
Werkzeug==3.0.1
yarl==1.9.2
